# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DocFlow AI Document Processing Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents with filtering
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [uploaded, processing, extracted, review_pending, approved, rejected, exported]
        - name: docType
          in: query
          schema:
            type: string
        - name: vendorId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Paginated document list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentPage"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Upload a new document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Document created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update document metadata or status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUpdate"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /documents/{id}/process:
    post:
      operationId: processDocument
      tags: [documents]
      summary: Trigger AI extraction on a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Processing started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}/reprocess:
    post:
      operationId: reprocessDocument
      tags: [documents]
      summary: Re-run extraction on a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reprocessing started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  # ── Extractions ─────────────────────────────────────────────────────────────
  /documents/{id}/extraction:
    get:
      operationId: getExtraction
      tags: [extractions]
      summary: Get AI extraction result for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Extraction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractionResult"
    patch:
      operationId: updateExtraction
      tags: [extractions]
      summary: Update corrected extraction fields
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtractionUpdate"
      responses:
        "200":
          description: Updated extraction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractionResult"

  # ── Validations ─────────────────────────────────────────────────────────────
...
  /documents/{id}/review:
    post:
      operationId: submitReview
      tags: [reviews]
      summary: Approve or reject a document review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewAction"
      responses:
        "200":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /review-queue:
    get:
      operationId: getReviewQueue
      tags: [reviews]
      summary: Get documents pending human review
      parameters:
        - name: priority
          in: query
          schema:
            type: string
            enum: [high, medium, low]
        - name: assignedTo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Review queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewQueueItem"

  # ── Workflows ───────────────────────────────────────────────────────────────
...
  /export-jobs:
    get:
      operationId: listExportJobs
      tags: [exports]
      summary: List export jobs
      responses:
        "200":
          description: Export jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExportJob"
    post:
      operationId: createExportJob
      tags: [exports]
      summary: Create an export job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportJobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportJob"
  /export-jobs/{id}:
    get:
      operationId: getExportJob
      tags: [exports]
      summary: Get export job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Export job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportJob"

  # ── Audit ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Document:
      type: object
      required: [id, fileName, fileType, status, docType, uploadSource, createdAt, updatedAt]
      properties:
        id:
          type: integer
        fileName:
          type: string
        fileType:
          type: string
        fileSizeBytes:
          type: ["integer", "null"]
        fileUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [uploaded, processing, extracted, review_pending, approved, rejected, exported]
        docType:
          type: string
          enum: [invoice, gst_bill, purchase_order, delivery_challan, shipment_slip, warehouse_doc, vendor_bill, receipt, other]
        uploadSource:
          type: string
          enum: [web_upload, email, whatsapp, api, bulk]
        vendorId:
          type: ["integer", "null"]
        vendorName:
          type: ["string", "null"]
        confidenceScore:
          type: ["number", "null"]
        isDuplicate:
          type: boolean
        duplicateOfId:
          type: ["integer", "null"]
        assignedTo:
          type: ["string", "null"]
        reviewComment:
          type: ["string", "null"]
        priority:
          type: string
          enum: [high, medium, low]
        tags:
          type: array
          items:
            type: string
        processingStartedAt:
          type: ["string", "null"]
        processingCompletedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    DocumentPage:
      type: object
      required: [documents, total, page, limit]
      properties:
        documents:
          type: array
          items:
            $ref: "#/components/schemas/Document"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    DocumentInput:
      type: object
      required: [fileName, fileType, docType, uploadSource]
      properties:
        fileName:
          type: string
        fileType:
          type: string
        fileSizeBytes:
          type: ["integer", "null"]
        fileUrl:
          type: ["string", "null"]
        docType:
          type: string
          enum: [invoice, gst_bill, purchase_order, delivery_challan, shipment_slip, warehouse_doc, vendor_bill, receipt, other]
        uploadSource:
          type: string
          enum: [web_upload, email, whatsapp, api, bulk]
        vendorId:
          type: ["integer", "null"]
        priority:
          type: string
          enum: [high, medium, low]
        tags:
          type: array
          items:
            type: string
    DocumentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [uploaded, processing, extracted, review_pending, approved, rejected, exported]
        docType:
          type: string
          enum: [invoice, gst_bill, purchase_order, delivery_challan, shipment_slip, warehouse_doc, vendor_bill, receipt, other]
        vendorId:
          type: ["integer", "null"]
        assignedTo:
          type: ["string", "null"]
        reviewComment:
          type: ["string", "null"]
        priority:
          type: string
          enum: [high, medium, low]
        tags:
          type: array
          items:
            type: string

    # ── Extraction ──────────────────────────────────────────────────────────────
    ExtractionResult:
      type: object
      required: [id, documentId, createdAt]
      properties:
        id:
          type: integer
        documentId:
          type: integer
        invoiceNumber:
          type: ["string", "null"]
        invoiceDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        poNumber:
          type: ["string", "null"]
        gstin:
          type: ["string", "null"]
        vendorGstin:
          type: ["string", "null"]
        vendorName:
          type: ["string", "null"]
        vendorAddress:
          type: ["string", "null"]
        vendorPhone:
          type: ["string", "null"]
        vendorEmail:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        customerAddress:
          type: ["string", "null"]
        shipmentId:
          type: ["string", "null"]
        warehouseId:
          type: ["string", "null"]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
        subtotal:
          type: ["number", "null"]
        taxAmount:
          type: ["number", "null"]
        cgst:
          type: ["number", "null"]
        sgst:
          type: ["number", "null"]
        igst:
          type: ["number", "null"]
        totalAmount:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        paymentTerms:
          type: ["string", "null"]
        bankDetails:
          type: ["string", "null"]
        rawText:
          type: ["string", "null"]
        fieldConfidences:
          type: object
          additionalProperties:
            type: number
        isHumanCorrected:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    LineItem:
      type: object
      properties:
        description:
          type: string
        hsnCode:
          type: ["string", "null"]
        quantity:
          type: ["number", "null"]
        unit:
          type: ["string", "null"]
        unitPrice:
          type: ["number", "null"]
        taxRate:
          type: ["number", "null"]
        amount:
          type: ["number", "null"]
    ExtractionUpdate:
      type: object
      properties:
        invoiceNumber:
          type: ["string", "null"]
        invoiceDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        poNumber:
          type: ["string", "null"]
        gstin:
          type: ["string", "null"]
        vendorGstin:
          type: ["string", "null"]
        vendorName:
          type: ["string", "null"]
        vendorAddress:
          type: ["string", "null"]
        vendorPhone:
          type: ["string", "null"]
        vendorEmail:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        customerAddress:
          type: ["string", "null"]
        shipmentId:
          type: ["string", "null"]
        warehouseId:
          type: ["string", "null"]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
        subtotal:
          type: ["number", "null"]
        taxAmount:
          type: ["number", "null"]
        cgst:
          type: ["number", "null"]
        sgst:
          type: ["number", "null"]
        igst:
          type: ["number", "null"]
        totalAmount:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        paymentTerms:
          type: ["string", "null"]

    # ── Validation ──────────────────────────────────────────────────────────────
...
    ReviewAction:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [approve, reject, request_changes]
        comment:
          type: ["string", "null"]
        assignedTo:
          type: ["string", "null"]
    ReviewQueueItem:
      type: object
      required: [document, validationCount, pendingSince]
      properties:
        document:
          $ref: "#/components/schemas/Document"
        validationCount:
          type: integer
        pendingSince:
          type: string
        assignedTo:
          type: ["string", "null"]

    # ── Workflow ────────────────────────────────────────────────────────────────
...
    ExportJob:
      type: object
      required: [id, exportFormat, status, createdAt]
      properties:
        id:
          type: integer
        exportFormat:
          type: string
          enum: [sap, tally, zoho, oracle, busy, excel, csv, json, google_sheets]
        status:
          type: string
          enum: [pending, processing, completed, failed]
        documentIds:
          type: array
          items:
            type: integer
        documentCount:
          type: integer
        downloadUrl:
          type: ["string", "null"]
        fieldMapping:
          type: object
          additionalProperties:
            type: string
        errorMessage:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    ExportJobInput:
      type: object
      required: [exportFormat, documentIds]
      properties:
        exportFormat:
          type: string
          enum: [sap, tally, zoho, oracle, busy, excel, csv, json, google_sheets]
        documentIds:
          type: array
          items:
            type: integer
        fieldMapping:
          type: object
          additionalProperties:
            type: string

    # ── Audit ────────────────────────────────────────────────────────────────────
...