# lib/api-zod/src/generated/types/documentPage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DocFlow AI Document Processing Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Document } from './document';

export interface DocumentPage {
  documents: Document[];
  total: number;
  page: number;
  limit: number;
}