# lib/api-zod/src/generated/types/extractionResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DocFlow AI Document Processing Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ExtractionResultFieldConfidences } from './extractionResultFieldConfidences';
import type { LineItem } from './lineItem';

export interface ExtractionResult {
  id: number;
  documentId: number;
  /** @nullable */
  invoiceNumber?: string | null;
  /** @nullable */
  invoiceDate?: string | null;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  poNumber?: string | null;
  /** @nullable */
  gstin?: string | null;
  /** @nullable */
  vendorGstin?: string | null;
  /** @nullable */
  vendorName?: string | null;
  /** @nullable */
  vendorAddress?: string | null;
  /** @nullable */
  vendorPhone?: string | null;
  /** @nullable */
  vendorEmail?: string | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerAddress?: string | null;
  /** @nullable */
  shipmentId?: string | null;
  /** @nullable */
  warehouseId?: string | null;
  lineItems?: LineItem[];
  /** @nullable */
  subtotal?: number | null;
  /** @nullable */
  taxAmount?: number | null;
  /** @nullable */
  cgst?: number | null;
  /** @nullable */
  sgst?: number | null;
  /** @nullable */
  igst?: number | null;
  /** @nullable */
  totalAmount?: number | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  paymentTerms?: string | null;
  /** @nullable */
  bankDetails?: string | null;
  /** @nullable */
  rawText?: string | null;
  fieldConfidences?: ExtractionResultFieldConfidences;
  isHumanCorrected?: boolean;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}